refactor(completion): rename cmdline_completions module to completion
